Split ExternalFunction out into its own module
